% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{center}
\alias{center}
\alias{center.standardized}
\title{Extract Values used for Standardization (scaled:center)}
\usage{
center(x, ...)

\method{center}{standardized}(x, ...)
}
\arguments{
\item{x}{object.}

\item{...}{forwarded to generic methods.}
}
\value{
Returns 'scaled:center' used for standardization
}
\description{
Generic method to extract 'centering' values used for
standardization (emprical first moment).
}
