% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{coef.foehnix}
\alias{coef.foehnix}
\title{Get Estimated Mixture Model Coefficients}
\usage{
\method{coef}{foehnix}(object, type = "parameter", ...)
}
\arguments{
\item{object}{\code{\link{foehnix}} mixture model object.}

\item{type}{character, either \code{"parameter"} \code{"coefficients"}, see
'Details' section.}

\item{...}{additional arguments, ignored.}
}
\value{
Returns a \code{coef.foehnix} object with the estimated
        coefficients.
}
\description{
Returns the estimated coefficients of a \code{\link{foehnix}} mixture model
for both, the components and the concomitant model (if in use).
}
\details{
Returns the coeffficients of the mixture model.
If \code{type = "parameter"} the parameters are returned of the
components are returned on the 'true' scale, namely:
\code{mu1} and \code{sigma1} (location and scale of component 1),
\code{mu2} and \code{sigma2} (location and scale of component 2),
and a set of coefficients from the concomitant model if one has been
specified. These coefficients are the coefficients from a (possibly
regularized) logistic regression model (see \code{\link{iwls_logit}}).

If \code{type = "coefficients"} the scale parameters are returned
on the log-scale (\code{logsd1}, \code{logsd2}) as used during
optimization.
}
\seealso{
\code{\link{foehnix}}, \code{\link{iwls_logit}}.
}
\author{
Reto Stauffer
}
