% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions.R
\name{ddff2uv}
\alias{ddff2uv}
\title{Convert Wind Speed and Wind Direction to u/v Wind Components}
\usage{
ddff2uv(dd, ff = NULL)
}
\arguments{
\item{dd}{only necessary if \code{ff} is a numeric vector.  Use NULL if
input \code{ff} is containing both variables (\code{ff}, \code{dd}), else
\code{dd} is a data vector containing the wind direction in degrees (0:
north, 90: east, ...)}

\item{ff}{\code{numeric}, \code{matrix}, \code{data.frame}, or
\code{zoo} object (see 'Details' section).}
}
\value{
Returns a \code{data.frame} or \code{zoo} object (depending on input \code{ff})
containing the \code{u} and \code{v} components
of the data. In addition, \code{rad} (mathematical representation
of wind direction in radiant) is returned.
}
\description{
Converts wind direction (\code{dd}) and wind speed (\code{ff})
information into \code{u}/\code{v} wind components (zonal and
meridional wind component).
}
\details{
Converts data from wind speed and direction into the zonal
and meridional wind components (\code{u}/\code{v}).

Different inputs are allowed:
\itemize{
   \item if \code{ff} is a matrix: requires to contains at least the
         two columns \code{"ff"} and \code{"dd"}.
   \item if \code{ff} is a \code{data.frame} or \code{zoo} object:
         requires to contains at least the
         two variables \code{"ff"} and \code{"dd"}.
   \item if \code{ff} is \code{numeric}: \code{dd} has to be provided
         in advance. \code{ff} and \code{dd} have to be of the same length
         or one has to be of length \code{1} (will be recycled).
}
}
\examples{
## Generate dd and ff variable
set.seed(0)
ff <- floor(abs(rnorm(20))*10)
dd <- sample(seq(0,359),20)
df <- data.frame('ff'=ff,'dd'=dd)

## Using with vectors 
print(head(ddff2uv('dd' = dd, 'ff' = ff)))

## Using with data.frame
print(head(ddff2uv(df)))

## Using with matrix
print(head(ddff2uv(as.matrix(df))))

## Use with zoo
library("zoo")
set.seed(100)
Sys.setenv("TZ" = "UTC")
data <- zoo(data.frame(dd = runif(20, 0, 360), ff = rnorm(20, .2, 2)^2),
            as.POSIXct("2019-01-01 12:00") + 0:19 * 3600)
data <- round(data,2)
head(data)
class(data)
## Calculate dd/ff
uv <- ddff2uv(data)
head(uv)
class(uv)
}
\seealso{
uv2ddff
}
\author{
Reto Stauffer
}
