% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demodatasets}
\alias{demodatasets}
\alias{ellboegen}
\alias{sattelberg}
\alias{viejas}
\alias{luckyfive}
\title{Meteorological Observation Data, Station Ellboegen}
\source{
Station Ellögen operated by the Department of Atmospheric and Cryospheric
Sciences (\url{http://acinn.uibk.ac.at}) of the University of Innsbruck.
The data is available under the
\href{https://creativecommons.org/licenses/by-sa/4.0/}{CC BY-SA 4.0 license}.

Station Sattelberg operated by the Department of Atmospheric and Cryospheric
Sciences (\url{http://acinn.uibk.ac.at}) of the University of Innsbruck.
The data is available under the
\href{https://creativecommons.org/licenses/by-sa/4.0/}{CC BY-SA 4.0 license}.

Station Viejas Casino and Resort operated by San Diego Gas and Electric and made freely
publicly available through Synoptic PBC at \url{https://synopticdata.com/}.

Station Lucky Five Ranch operated by San Diego Gas and Electric and made freely
publicly available through Synoptic PBC at \url{https://synopticdata.com/}.
}
\usage{
# Combined data set Tyrol (A).
# Valley station: Ellbögen; crest station: Sattelberg
demodata("tyrol")

# Combined data set California (CA, USA).
# Valley station: Viejas; crest station: Lucky Five Ranch
demodata("tyrol")

# Hourly observations station Ellögen (Tyrol, Austria).
# Serves as "valley" or "target" station for the "tyrol" data set.
data("ellboegen")

# Hourly observations station Sattelberg (Tyrol, Austria).
# Serves as "crest" station for the "tyrol" data set.
data("sattelberg")

# Hourly observations station Viejas Casino and Resort (California, U.S.A.).
# Serves as "valley" or "target" station for the "california data set.
data("viejas")

# Hourly observations station Lucky Five Ranch (California, U.S.A.).
# Serves as "crest" station for the "california" data set.
data("luckyfive")
}
\description{
The \code{\link{foehnix}} package comes with two data sets of meteorological
observations for two sites in Tyrol, Austria.  One station is located near
Ellboegen, a small town close to Innsbruck, the second station (Sattelberg)
is located south of Ellboegen close to the border between Austria and Italy
on top of a mountain close to the crest of the European Alps and is used as
'crest station' (see \code{\link{sattelberg}}).
}
\details{
#TODO migrate the descriptions (commented, see below) into this combined
manual page.


Ellboegen:
\itemize{
    \item Location: 11.42889 E/47.18694 N
    \item Altitude: 1080 meters above mean sea level
}

\code{zoo} time series object with an hourly temporal resolution (time zone: UTC)
including the following variables:
\itemize{
   \item \code{dd}: wind direction in degrees, meteorological format
       (0/360 correspond to 'wind from north', 90 to 'wind from east',
       180 'wind from south', and 270 'wind from west')
   \item \code{ff}: wind speed in meters per second
   \item \code{p}: station pressure in hectopascal
   \item \code{rh}: relative humidity in percent
   \item \code{t}: dry air temperature in degrees Celsius
}
}
\seealso{
\code{\link{sattelberg}}, \code{\link{demodata}},
         \code{\link{viejas}}, \code{\link{luckyfive}}.
}
\author{
Reto Stauffer, Deborah Detka
}
\keyword{data}
