% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{edf}
\alias{edf}
\title{Returns Effective Degrees of Freedom}
\usage{
edf(object, ...)
}
\arguments{
\item{object}{the object from which the effective degrees
of freedom should be returned.}

\item{...}{forwarded to class specific edf methods.}
}
\description{
Function which returns the effective degrees of freedom.
}
