% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix.R
\name{fitted.foehnix}
\alias{fitted.foehnix}
\title{Returns Fitted Values of foehnix Mixture Models}
\usage{
\method{fitted}{foehnix}(object, which = "probability", ...)
}
\arguments{
\item{object}{a \code{\link{foehnix}} model object.}

\item{which}{what to get as return. Can a character string, one of
\code{"probability"} or \code{"flag"}, or a vector with both
(\code{c("probability", "flag")}) to get both values. Alternatively
integers can be used (\code{1}, \code{2}, or \code{c(1,2)}).}

\item{...}{additional arguments, ignored.}
}
\value{
Returns a univariate or multivariate \code{zoo} object.
}
\description{
Extracts fitted probabilities and/or flags from a \code{\link{foehnix}}
mixture model.
}
\author{
Reto Stauffer
}
