% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix.R
\name{foehnix.control}
\alias{foehnix.control}
\title{foehnix Two-Component Mixture-Model Control Object}
\usage{
foehnix.control(family, switch, left = -Inf, right = Inf,
  truncated = FALSE, standardize = TRUE, maxit = 100L, tol = 1e-08,
  force.inflate = FALSE, verbose = TRUE, glmnet.control = NULL, ...)
}
\arguments{
\item{family}{character specifying the distribution of the components in the
mixture model. Allowed: \code{"gaussian"} and \code{"logistic"}.  For
experts: custom \code{foehnix.family} objects can be provided as well.}

\item{switch}{logical whether or not the two components should be switched.
By default (\code{switch = FALSE}) the component which shows
higher values of \code{y} is assumed to be the foehn cluster!
Depending on what your covariate is you might need to switch
the clusters (by setting \code{switch = TRUE}).}

\item{left}{default is \code{-Inf}, left censoring or truncation point.  See
also input \code{right} and input \code{truncated}. Can be set to any finite
numeric value.}

\item{right}{default is \code{Inf}, right censoring or truncation point.  See
also input \code{left} and input \code{truncated}. Can be set to any finite
numeric value.}

\item{truncated}{logical. If set to \code{TRUE} truncation is used instead of
\code{censored}. This only affects the \code{\link{foehnix}} model estimate
if input \code{left} and/or input \code{right} are specified.}

\item{standardize}{logical flag, default is \code{TRUE}. Defines whether or
not the model matrix for the concomitant model should be standardized for
model estimation. Recommended.}

\item{maxit}{control argument for the iterative solvers. Default is
\code{100L}, the maximum number of iterations for the EM algorithm and the
IWLS backfitting algorithm for the concomitant model.  If a vector of length
two is provided the first value is used for the EM algorithm, the second for
the IWLS backfitting.}

\item{tol}{similar as for \code{maxit}. Used to identify convergence of the
iterative solvers. Default is \code{1e-8}, if two values are given the first
will be used for the EM algorithm, the second one for the IWLS backfitting
procedure. If set to \code{-Inf} \code{maxit} will be used as stopping
criteria.}

\item{force.inflate}{logical, default is \code{FALSE}. \code{\link{foehnix}} creates
a strictly regular time series object by inflating the data set using the
smallest time interval in the data set. If the inflation rate is larger than
2 the script will stop except the user forces inflation by specifying
\code{force.inflate = TRUE}. See 'Details' section for more information.}

\item{verbose}{logical, if set to \code{FALSE} output is suppressed.}

\item{...}{currently sent to hell.}
}
\description{
Used to control the \code{\link{foehnix}} mixture models.
}
\details{
\code{\link{foehnix}} models are based on time series objects. 
For some methods (e.g., to create nice and easy to read time series plots and
count statistics)
\code{\link{foehnix}} inflates the time series object using the smallest time
interval in the data set. This can, possibly, yield very large data sets. Thus,
\code{\link{foehnix}} is pre-calculating the inflation rate, the fraction between
the length of the inflated data set versus the length of the data set provided by
the user. If this inflation rate exceeds 2 the script will raise an error!

In this case the user should make sure that the time series object provided
is proper before continuing. A possible scenario: a user is performing foehn diagnosis
using 5 years of data from one station with 10 minute observations. This yields
(neglecting leap years) \code{5 * 365 * 144 = 262.800} observations. Imagine that there
is one incorrect observation reported one second after one of the regular
10 minute records. The smallest time increment would thus be 1 second. This would
yield an inflated time series object with a total record length of
\code{5 * 365 * 86.400 = 157.680.000}. Even if only filled with missing values
(\code{NA}) this will be extremely memory demanding. To avoid this action
\code{\link{foehnix}} will stop in such situations.

However, the user is allowed to overrule this condition by setting the
\code{force.inflate} option to \code{TRUE}.
}
\seealso{
\code{\link{foehnix}}, \code{\link{foehnix.family}}.
}
\author{
Reto Stauffer
}
