% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{foehnix.family}
\alias{foehnix.family}
\alias{print.foehnix.family}
\alias{is.truncated}
\alias{has.left}
\alias{has.right}
\alias{foehnix_logistic}
\alias{foehnix_clogistic}
\alias{foehnix_tlogistic}
\alias{foehnix_gaussian}
\alias{foehnix_cgaussian}
\alias{foehnix_tgaussian}
\title{foehnix Families for Two-Component Mixture Models}
\usage{
\method{print}{foehnix.family}(x, ...)

is.truncated(x, ...)

has.left(x, ...)

has.right(x, ...)

foehnix_logistic()

foehnix_clogistic(left = -Inf, right = Inf)

foehnix_tlogistic(left = -Inf, right = Inf)

foehnix_gaussian()

foehnix_cgaussian(left = -Inf, right = Inf)

foehnix_tgaussian(left = -Inf, right = Inf)
}
\arguments{
\item{x}{\code{foehnix.family} object.}

\item{...}{additional arguments, ignored.}

\item{left}{left censoring or truncation point (if the family object
provides this option).}

\item{right}{right censoring or truncation point (if the family object
provides this option).}
}
\description{
The foehnix mixture models are based on a set of families provided with this
package. Currently, the package provides a two-component Gaussian and a
two-component logistic mixture model and their truncated and censored
versions.  Custom \code{foehn.family} objects could also be plugged in (see
'Details' section).
}
\details{
The method \code{\link{foehnix}} allows to specify a \code{family} input
argument which has to be either \code{"gaussian"} (the default) or
\code{"logistic"} for now.  If finite arguments for \code{left} and/or
\code{right} are set as well, a censored Gaussian/logistic mixture model
will be estimated (or truncated, if \code{truncated = TRUE}; see
\code{\link{foehnix}} and \code{\link{foehnix.control}}).

However, feel free to develop custom family objects if needed: if a
\code{foehnix.family} object is provided on \code{family} when calling
\code{foehnix} this custom object will be used. For example:

\itemize{
    \item \code{fam <- foehnix:::foehnix_cgaussian(left = 0)}
    \item \code{mod <- foehnix(dt ~ ff + dd + rh, data = data, family = fam)}
} 

Each `foehn.family` object consists of a set of functions:
\itemize{
    \item optional arguments not listed here (stored inside the
          object environment when specified on initialization,
    \item \code{left} and \code{right} used for the censored and truncated
          Gaussian/logistic families are such examples).
    \item \code{name}: character, name of the family object.
    \item \code{d}: density function of the mixture distribution.
    \item \code{p}: distribution function of the mixture distribution.
    \item \code{r}: not required for optimization but nice for testing:
          returns random numbers from the mixed distribution to simulate data.
    \item \code{loglik} function returning the log-likelihood sum,
          used for model estimation (EM algorithm).
    \item \code{posterior} returns the (updated) posterior probabilities,
          used for model estimation (EM algorithm).
    \item \code{theta} returns the parameters of the distributions of the
          components of the mixture models. Used for model
          estimation (EM algorithm).
}
Examples are: `foehnix_gaussian`, `foehnix_cgaussian`, `foehnix_tgaussian`,
`foehnix_logistic`, `foehnix_clogistic`, and `foehnix_tlogistic` in
`R/families.R`.
}
\seealso{
\code{\link{foehnix}}, \code{\link{foehnix.control}}.
}
\author{
Reto Stauffer
}
