\name{foehnix.reg.fit}
\alias{foehnix.reg.fit}
\title{Fitting Regularized foehnix Mixture Model With Concomitant Model.}
\usage{
    foehnix.reg.fit(...)
}
\description{
    Fitting method for two-component foehnix mixture model with
    regularization. Typically not called directly, interfaced via \code{\link{foehnix}}.

    TODO: Method not yet implemented, as soon as the method itself
    has been written: update manual page!
}
\arguments{
    \item{\dots}{additional arguments, unused.}
}
\seealso{
    \code{\link{foehnix}},
    \code{\link{foehnix.control}},
    \code{\link{foehnix.noconcomitant.fit}},
    \code{\link{foehnix.reg.fit}},
    \code{\link{iwls_logit}}.
}
