% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_glmnet.R
\name{foehnix_glmnet}
\alias{foehnix_glmnet}
\title{Wrapper Function Around glmnet Logistic Regression}
\usage{
foehnix_glmnet(y, x, arg)
}
\arguments{
\item{y}{response vector (binary)}

\item{x}{design matrix. If a column \code{"(Intercept)"} exists it will be
dropped (as we force glmnet to estimate an intercept).}

\item{arg}{list of arguments forwarded to \code{glmnet::glmnet}.}
}
\value{
Returns an object of class \code{ccmodel} for \code{\link[foehnix]{foehnix}}
models.
}
\description{
The \code{\link[foehnix]{foehnix}} function allows to estimate
a regularized logistic regression model for the concomitants.
In case \code{\link[foehnix]{glmnet.control}} is provided the
EM algorithm uses this function to estimate the regression coefficients
of the concomitant model. \code{\link[foehnix]{glmnet.control}} allows
to specify options forwarded to \code{glmnet::glmnet} except 
family (\code{"binomial"}) and intercept (\code{TRUE}).
Depending on \code{\link[foehnix]{glmnet.control}} the AIC, BIC, or
log-likelihood criterion will be used.
}
\author{
Reto Stauffer
}
