% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_glmnet.R
\name{glmnet.control}
\alias{glmnet.control}
\alias{print.glmnet.control}
\title{Control Object to Allow for Regularized Concomitant Models}
\usage{
glmnet.control(min = "AIC", ...)

\method{print}{glmnet.control}(x, ...)
}
\arguments{
\item{min}{a character, either \code{"AIC"}, \code{"BIC"}, or \code{"loglik"}.
Default is \code{"AIC"}.}

\item{...}{additional arguments forwarded to \code{glmnet::glmnet}.}

\item{x}{an object of class \code{glmnet.control} (print method).}
}
\description{
\code{foehnix} allows to estimate regularized concomitant models
based on the R package glmnet. The \code{\link[foehnix]{glmnet.control}}
object controls, if specified, the glmnet input arguments.
}
\details{
Note that the two input arguments \code{intercept} and
\code{family} (to function \code{glmnet}) cannot be overruled by the
foehnix user. In any case a binomial logit model with intercept will
be estimated.
}
\author{
Reto Stauffer
}
