% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{IGN}
\alias{IGN}
\title{Returns the Ignorance of the Object}
\usage{
IGN(object, ...)
}
\arguments{
\item{object}{the object to be evaluated.}

\item{...}{forwarded to generic methods.}
}
\description{
Mmethod which returns the ignorance of the \code{object}.
The ignorance is the mean negative log-likelihood.
}
\seealso{
foehnix()
}
