% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{image.foehnix}
\alias{image.foehnix}
\title{foehnix Image Plot - Hoevmoeller Diagram}
\usage{
\method{image}{foehnix}(x, FUN = "freq", contours = FALSE,
  deltat = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link{foehnix}}.}

\item{FUN}{character string or a custom aggregation function. See 'Details'
section for more information.}

\item{contours}{logical \code{TRUE} or \code{FALSE}, whether or not the
concours should be added.}

\item{deltat}{integer, interval in seconds for the time of day axis. Has to be
a fraction of 86400 (24 hours in seconds). It \code{NuLL} (default) the
interval of the time series object will be used.}

\item{...}{additional arguments, currently unused.}
}
\description{
The default \code{\link{image}} plot of a \code{\link{foehnix}} object
is a Hoevmoeller diagram.
}
\details{
Plotting a Hoevmoeller diagram based on the \code{\link{zoo}} time
series object of the \code{\link{foehnix}} classification. Different plot
types are available. The default functions (see list below) use \code{na.rm = TRUE}.

Input \code{FUN} can be one of the following character strings:
\itemize{
   \item \code{mean}: mean probability.
   \item \code{freq}: plotting frequencies.
   \item \code{occ}: plotting occurance of foehn (where probability \code{>= 0.5}).
   \item \code{noocc}: contrary to \code{occ}: occurance of no foehn (probability \code{< 0.5}).
}

\code{FUN} can also be a custom function used for time series aggregation
(see \code{\link{aggregate.zoo}}).
}
