% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image}
\alias{image}
\alias{image.foehnix}
\title{foehnix Image Plot - Hovmoeller Diagram}
\usage{
\method{image}{foehnix}(x, FUN = "freq", deltat = NULL, deltad = 7L,
  col = rev(gray.colors(20)), contours = FALSE,
  contour.col = "black", ...)
}
\arguments{
\item{x}{object of class \code{\link{foehnix}}.}

\item{FUN}{character string or a custom aggregation function. See 'Details'
section for more information.}

\item{deltat}{integer, interval in seconds for the time of day axis. Has to be
a fraction of 86400 (24 hours in seconds). It \code{NuLL} (default) the
interval of the time series object will be used.}

\item{deltad}{integer, similar to \code{deltat}, the interval in days for the
grid on the x-axis. Default is \code{7L} (aggregate to weekly values).}

\item{col}{vector of colors forwarded to \code{image.default}. By default
a gray scale color map is used.}

\item{contours}{logical \code{TRUE} or \code{FALSE}, whether or not the
concours should be added.}

\item{contour.col}{color for the contour lines, only used if \code{contours = TRUE}.}

\item{...}{additional arguments (see 'Details' section).}
}
\description{
The default \code{\link{image}} plot of a \code{\link{foehnix}} object
is a Hovmoeller diagram.
}
\details{
Plotting a Hovmoeller diagram based on the \code{\link{zoo}} time
series object of the \code{\link{foehnix}} classification. Different plot
types are available. The default functions (see list below) use \code{na.rm = TRUE}.

Input \code{FUN} can be one of the following character strings:
\itemize{
   \item \code{freq}: plotting frequencies (default).
   \item \code{mean}: mean probability.
   \item \code{occ}: plotting occurrence of foehn (where probability \code{>= 0.5}).
   \item \code{noocc}: contrary to \code{occ}: occurrence of no foehn (probability \code{< 0.5}).
}

\code{FUN} can also be a custom function used for time series aggregation
(see \code{\link{aggregate.zoo}}).

Additional arguments which can be set:
\itemize{
     \item \code{xlim}: limits of abscissa. Numeric vector of length 2 with
           Julian days (0 - 365). If the values are decreasing (e.g., \code{xlim = c(300, 100)})
           the abscissa will be adjusted to show continuous data around new years eve.
     \item \code{ylim}: limits of ordinate. Numeric vector of length 2 with
           seconds (seconds of the day; 0 = 00:00:00 UTC, 86400 = 24:00:00 UTC).
     \item \code{xlab}, \code{ylab}, \code{main}: axis labels and title of the plot.
}
}
