% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{is.standardized}
\alias{is.standardized}
\title{Check if Matrix is Standardized}
\usage{
is.standardized(x, ...)
}
\arguments{
\item{x}{a matrix or standardized matrix.}

\item{...}{ignored.}
}
\description{
Returns \code{TRUE}
if input \code{x} is a standardized matrix, else \code{FALSE}.
}
\seealso{
\code{\link[foehnix]{standardize}}.
}
