\name{phoeton-package}
\alias{phoeton-package}
\alias{phoeton}
\docType{package}
\title{
\packageTitle{phoeton}
}
\description{
\packageDescription{phoeton}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{phoeton}
\packageIndices{phoeton}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{phoeton}

Maintainer: \packageMaintainer{phoeton}
}
