% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.foehnix.R
\name{plot.foehnix}
\alias{plot.foehnix}
\title{foehnix Model Assessment Plots}
\usage{
\method{plot}{foehnix}(x, which = NULL, log = TRUE, ..., ask = TRUE)
}
\arguments{
\item{x}{a \code{\link{foehnix}} mixture model object.}

\item{which}{\code{NULL} (default), character, character string,
integer, or numeric. Allowed characters: \code{loglik} (\code{1}),
\code{loglikcontribution} (\code{2}), \code{coef} (\code{3}),
\code{hist} (\code{4}), and \code{posterior} (\code{5}).}

\item{log}{logical, if \code{TRUE} the x-axis for \code{loglik}, \code{loglikcontribution}
and \code{coef} is shown on the log scale.}

\item{...}{additional arguments, unused.}

\item{ask}{boolean, default is \code{TRUE}. User will be asked to show the
next figure if multiple figures are requested. Can be set to \code{FALSE}
to overwrite the default.}
}
\description{
Visual representation \code{\link{foehnix}} model optimization.
}
\details{
There are currently three different plot types.
\itemize{
    \item \code{"loglik"} shows the log-likelihood sum path trough
        the iterations of the EM algorithm for parameter estimation.
    \item \code{"loglikcontribution"} shows the log-likelihood
        contribution (initial value subtracted; all paths start
        with \code{0}).
    \item \code{coef} shows the development of the (standardized)
        coefficients during EM optimization. Parameters of the
        components are shown on the real scale, the coefficients
        of the concomitant model (if used) are shown on the
        standardized scale.
    \item \code{hist} plots empirical histograms of the main variable
        for the two clusters (split at posterior probability >= 0.5).
        In addition, the estimated parametric clusters are shown.
    \item \code{posterior} creates a histogram of the posterior probabilities.
        point masses at \code{0.0} and \code{1.0} indicate sharp separation
        of the clusters (posterior probabilities close to \code{0} or \code{1}).
        The two additional input arguments \code{breaks} (numeric vector)
        and \code{freq} (logical) will be forwarded to \code{hist(...)}.
}
}
\author{
Reto Stauffer
}
