% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix.R
\name{predict.foehnix}
\alias{predict.foehnix}
\title{Predict Method for foehnix Mixture Models}
\usage{
\method{predict}{foehnix}(object, newdata = NULL, type = "response", ...)
}
\arguments{
\item{object}{a \code{\link[foehnix]{foehnix}} mixture model object.}

\item{newdata}{if \code{NULL} (default) the prediction of the
underlying training data set will be returned
(see also \code{\link[foehnix]{fitted.foehnix}}). Else \code{newdata}
has to be a \code{zoo} object providing the required
variables which have been used for model fitting
and filtering (see \code{\link[foehnix]{foehnix}}).}

\item{type}{character, one of \code{"response"} (default),
\code{"all"}.}

\item{...}{additional arguments, ignored.}
}
\value{
Returns a \code{zoo} object with foehn probabilities
and (if \code{type = "all"}) additional information. See 'Details'
section for more information.
}
\description{
Some details.
}
\details{
Used for prediction (perform foehn diagnosis given the
estimated parameters on a new data set (\code{newdata}). If no
new data set is provided (\code{newdata = NULL}) the prediction
is made on the internal data set, the data set which has been
used to train the \code{\link[foehnix]{foehnix}} mixture model.
If a new data set is provided (\code{newdata}) the foehn diagnosis
will be performed on this new data set, e.g., based on a set
of new observations when using \code{\link[foehnix]{foehnix}} for operational
near real time foehn diagnosis.

Note that \code{newdata} has to be a \code{zoo} object containing
the required information to perform the \code{\link[foehnix]{foehnix}} diagnosis,
namely the variables used for classification (see
\code{\link[foehnix]{formula.foehnix}} plus the ones used to filter the data
(see \code{\link[foehnix]{foehnix}} input argument \code{filter}).

Usually \code{type = "response"} is used which returns the foehn
probabilities. If \code{type = "all"} a set of additional values will
be returned, namely:
\itemize{
    \item \code{density1} density of the first component of the mixture model.
    \item \code{density2} density of the second component (foehn component) of the
          mixture model.
    \item \code{ccmodel} probability from the concomitant model.
}
Note that the foehn probability is simply given by:
\itemize{
    \item ccmodel * density2 / ((1 - ccmodel) * density1 + ccmodel * density2)
}
}
\author{
Reto Stauffer
}
