% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_check}
\alias{shiny_check}
\title{Shiny Application to Check foehnix Classification (Time Series Plots)}
\usage{
shiny_check(x, ...)
}
\arguments{
\item{x}{a \code{\link{foehnix}} mixture model object.}

\item{...}{additional arguments, not in use.}
}
\description{
TODO: Neither well coded, nor tested, nor documented.
}
\author{
Reto Stauffer
}
