% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot.control}
\alias{tsplot.control}
\alias{get.tsplot.control}
\title{foehnix Time Series Plot Config}
\usage{
tsplot.control(...)

get.tsplot.control(x, var, property)
}
\arguments{
\item{...}{a set of named inputs to overwrite the defaults.
see 'Details' section.}

\item{x}{a \code{\link[foehnix]{tsplot.control}} object.}

\item{var}{used when calling the \code{\link[foehnix]{get.tsplot.control}}
function. Name of the (original!) variable name}

\item{property}{the property which should be returned by
\code{\link[foehnix]{get.tsplot.control}}.}
}
\value{
Returns an object of class \code{c("tsplot.control", "data.frame")}
with the specification for the time series plot.
}
\description{
The foehnix package comes with a default time series plotting
function (see \code{\link[foehnix]{tsplot}}) with a pre-specified
behavior regarding variable names, colors, and labels.
The \code{\link[foehnix]{tsplot}} function, however, allows to
set some custom specifications, e.g., if the names of your
observations (variable name of the observations in the time
series object used to train the \code{\link[foehnix]{foehnix}}
model) are different, if different labels are required, or
if you do not like our pretty colors! This function returns
a control object for the time series plots for customization.
}
\details{
By default the \code{\link[foehnix]{tsplot}} function
    expects that the variable names are called
\itemize{
     \item \code{t}: dry air temperature (degrees Celsuis)
     \item \code{rh}: relative humidity in percent
     \item \code{diff_t}: temperature difference to a nearby
            crest station
     \item \code{dd}: wind direction in (meteorological) degrees
     \item \code{ff}: wind speed in meters per second
     \item \code{ffx}: gust speed in meters per second
}
If \code{\link[foehnix]{tsplot}} can find these variables,
it will plot the corresponding observations, label the plots,
and uses a set of default colors.
The \code{\link[foehnix]{tsplot.control}} function allows
to overrule the defaults by specifying custom values, e.g.:
imagine that your wind speed variable is not \code{ff} but
\code{windspd}. You can easily tell \code{\link[foehnix]{tsplot}}
that it has to use this custom name by calling \code{\link[foehnix]{tsplot}}
as follows:
\itemize{
     \item \code{tsplot(x, ff = "windspd")}
}
If your wind speed variable is not even in meters per second but knots, and
you dislike our default color, you can also provide a list to overwrite the
default \code{name}, default \code{color} and default \code{label} by calling:
\itemize{
     \item \code{tsplot(x, ff = list(name = "windspd", color = "#0000ff", label = "wind speed in knots")}
}
In the same way it can be used to overrule the defaults for all other
variables used in the time series plotting function (see list above).
}
\author{
Reto Stauffer
}
