% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot_add_temp}
\alias{tsplot_add_temp}
\alias{tsplot_add_tempdiff}
\alias{tsplot_add_wind}
\title{Helper functions to add data to time series plots}
\usage{
tsplot_add_temp(x, control, prob_boxes)

tsplot_add_tempdiff(tmp, control, prob_boxes)

tsplot_add_wind(x, control, prob_boxes, crest = FALSE)
}
\arguments{
\item{x}{\code{zoo} object with the data}

\item{control}{\code{\link[foehnix]{tsplot.control}} object}

\item{prob_boxes}{object as returned from \code{tsplot_calc_prob_boxes}}
}
\description{
Used inside \code{\link[foehnix]{tsplot}}.
}
\author{
Reto Stauffer

Reto Stauffer

Reto Stauffer
}
