% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot_add_boxes}
\alias{tsplot_add_boxes}
\title{Helper function to add the gray boxes (background)}
\usage{
tsplot_add_boxes(x, col = "gray90")
}
\description{
Helper function to add the gray boxes (background)
}
