% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot_add_foehn}
\alias{tsplot_add_foehn}
\title{Add foehn probabilities to time series plot}
\usage{
tsplot_add_foehn(x, control, prob_boxes, xtra = NULL,
  xtra_names = NULL, ...)
}
\description{
Add foehn probabilities to time series plot
}
