% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot_add_midnight_lines}
\alias{tsplot_add_midnight_lines}
\title{Helper function to add vertical lines (midnight)}
\usage{
tsplot_add_midnight_lines(x)
}
\description{
Helper function to add vertical lines (midnight)
}
