% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions.R
\name{uv2ddff}
\alias{uv2ddff}
\title{Convert U and V Wind Components to Wind Speed and Direction}
\usage{
uv2ddff(u, v = NULL, rad = FALSE)
}
\arguments{
\item{u}{\code{numeric} vector with U components or a 
\code{data.frame} or \code{zoo} object containing
a column named \code{u} and a column named \code{v}.}

\item{v}{\code{NULL} (default) or \code{numeric} vector.
Iv input \code{u} is a single vector \code{v} has to be given.}

\item{rad}{\code{logical}, default is \code{FALSE}. Returns wind
direction in radiant rather than in degrees.}
}
\value{
Returns a data.frame with two columns named \code{dd}
   and \code{ff} containing wind speed (same physical unit as
   input \code{u}/\code{v}) and wind direction. Wind direction
   is either in meteorological degrees (0 from North, from 90 East,
   180 from South, and 270 from West) or in mathematical radiant
   if input \code{rad = TRUE}.
}
\description{
Takes U/V wind components as input and returns wind speed
and meteorological wind direction.
}
\examples{
## Generate data.frame with u/v components for all 4 main wind directions
data <- data.frame( name = c("N","E","S","W"),
                    u    = c( 0, -1 , 0,  1 ),
                    v    = c(-1,  0,  1,  0 ) )
## Use data.frame input
ddff <- uv2ddff( data )
cbind( data, ddff )
## Use u/v components as two separate vector inputs
ddff <- uv2ddff( data$u, data$v )
cbind( data, ddff )
## Radiant
ddff <- uv2ddff( data, rad = TRUE )
cbind( data, ddff )
}
\seealso{
ddff2uv
}
