% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{viejas}
\alias{viejas}
\title{Meteorological Observation Data, Station Viejas}
\source{
Station operated by San Diego Gas and Electric and made freely
publicly available through Synoptic PBC at \url{https://synopticdata.com/}.

Station operated by San Diego Gas and Electric and made freely
publicly available through Synoptic PBC at \url{https://synopticdata.com/}.
}
\usage{
data("viejas")
}
\description{
The \code{\link{foehnix}} package comes with two data sets of meteorological
observations for two sites in south California, USA. One station is located
next to the Viejas Casino and Resort located in Alpine, CA (valley station),
the second station called 'Lucky Five Ranch' is located 22km (14mi)
northwest of the Viejas Casino close to the main ridge of the Sierra Nevada
mountain range and is used as the crest station.
}
\details{
Viejas Casino and Resort:
\itemize{
    \item Location: -116.70437 E/32.84559 N
    \item Altitude: 715 meters above mean sea level
}

\code{zoo} time series object with an hourly temporal resolution (time zone: UTC)
including the following variables:
\itemize{
   \item \code{air_temp}: dry air temperature in degrees Celsius
   \item \code{relative_humidity}: relative humidity in percent
   \item \code{wind_speed}: wind speed in meters per second
   \item \code{wind_direction}: wind direction in degrees, meteorological format
       (0/360 correspond to 'wind from north', 90 to 'wind from east',
       180 'wind from south', and 270 'wind from west')
   \item \code{wind_gust}: wind gust speed in meters per second
}

Lucky Five Ranch:
\itemize{
    \item Location: -116.528 E/32.9331 N
    \item Altitude: 1445 meters above mean sea level
}

\code{zoo} time series object with an hourly temporal resolution (time zone: UTC)
including the following variables:
\itemize{
   \item \code{air_temp}: dry air temperature in degrees Celsius
   \item \code{relative_humidity}: relative humidity in percent
   \item \code{wind_speed}: wind speed in meters per second
   \item \code{wind_direction}: wind direction in degrees, meteorological format
       (0/360 correspond to 'wind from north', 90 to 'wind from east',
       180 'wind from south', and 270 'wind from west')
   \item \code{wind_gust}: wind gust speed in meters per second
}
}
\seealso{
\code{\link{sattelberg}}, \code{\link{demodata}},
         \code{\link{viejas}}, \code{\link{luckyfive}}.

\code{\link{sattelberg}}, \code{\link{demodata}},
         \code{\link{viejas}}, \code{\link{luckyfive}}.
}
\author{
Reto Stauffer, Georg J. Mayr
Meteorological Observation Data, Lucky Five Ranch

The \code{\link{foehnix}} package comes with two data sets of meteorological
observations for two sites in south California, USA. One station is located
next to the Viejas Casino and Resort located in Alpine, CA (valley station),
the second station called 'Lucky Five Ranch' is located 22km (14mi)
northwest of the Viejas Casino close to the main ridge of the Sierra Nevada
mountain range and is used as the crest station.

Reto Stauffer, Georg J. Mayr
}
\keyword{data}
\keyword{luckyfive}
\keyword{viejas}
