% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose_add_windsector}
\alias{windrose_add_windsector}
\title{Adding Wind Sector for Windrose Plots}
\usage{
windrose_add_windsector(x, ff, col, offset = 0.02)
}
\arguments{
\item{x}{list containing vectors of length 2 or a matrix with two
rows. If the list is named, or the matrix has row names,
the names will be used to label the segments.}

\item{ff}{wind speed (in case of plot type "density", or density in 
case of plot type "histogram". Used to draw the polygon and
to plot the labels (if needed).}

\item{col}{color of the wind sector.}

\item{offset}{offset for text adjustment.}
}
\description{
The windrose plot allows to specify one or multiple
segments to be highlighted on the plot. This is a
helper function to place the segments.
}
\author{
Reto Stauffer
}
