% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose_get_cols}
\alias{windrose_get_cols}
\title{Get Color Matrix and Legend for Windrose Plot}
\usage{
windrose_get_cols(x, col, p = 1, ncol = 50L)
}
\arguments{
\item{x}{2-D matrix with counts (see \code{\link{windrose_get_counts}}).}

\item{col}{single hex color, or a vector of hex colors. Alpha channel
will be removed.}

\item{p}{numeric value, power parameter for non-linear color transformation.}

\item{ncol}{integer, if \code{col} is a single value the color will be repeated
\code{ncol} times. Default is \code{50L}.}
}
\value{
Returns a list with two elements: \code{colormatrix} of the same
dimension as \code{x} with colors for the different bins based on the
counts, and \code{legend} with levels and colors for the color legend of the
plot.
}
\description{
Get Color Matrix and Legend for Windrose Plot
}
