% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose_get_counts}
\alias{windrose_get_counts}
\title{Get Count Matrix for Windrose Plot}
\usage{
windrose_get_counts(
  x,
  dd.breaks = seq(0, 360, by = 30),
  ff.breaks = pretty(x$ff)
)
}
\arguments{
\item{x}{data object of type zoo or data.frame. Needs to contain
at least the two columns \code{dd} (meteorological wid direction
in degrees, \code{]0, 360[}) and \code{ff} with wind speed
(range >= 0).}

\item{dd.breaks}{numeric vector with breaks along wind direction.
Default is \code{seq(0, 360, by = 30)}.}

\item{ff.breaks}{numeric vector with breaks along wind speed.
default is \code{pretty(x$ff)}.}
}
\value{
Returns a matrix of dimension \code{length(dd.breaks) x length(ff.breaks)}
        with counts \code{>= 0}.
}
\description{
The \code{\link{windrose}} method provides one windrose/wind count
plot type. This function returns a matrix with counts for different
multivariate bins (binning along wind direction dd and wind speed ff).
}
