% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{windsector_convert}
\alias{windsector_convert}
\title{Convert input "windsector" to list if needed.}
\usage{
windsector_convert(x)
}
\arguments{
\item{x}{\code{NULL}, \code{list}, \code{matrix}, or \code{data.frame}.
see 'Details' for more information.}
}
\value{
Returns a \code{list} object with the windsector information as used
by the different functions and methods of the \code{foehnix} package.
}
\description{
Converts user inputs to list. Wind sector information is used
to highlight specific wind sectors on the plots (e.g.,
\code{\link[foehnix]{tsplot}}, \code{\link[foehnix]{windrose}}).
}
\details{

}
\examples{
# No wind sector (NULL) simply returns NULL
foehnix:::windsector_convert(NULL)

# Input can also be:
# - unnamed list
# - a matrix
# - a data.frame without row names
foehnix:::windsector_convert(matrix(c(10, 30, 90, 140), byrow = TRUE, ncol = 2))
foehnix:::windsector_convert(list(c(10, 30), c(90, 140)))
foehnix:::windsector_convert(data.frame(from = c(30, 90), to = c(30, 140)))

# Or named objects can be used
# - named list
# - matrix with rownames
# - data.frame with rownames
# If names are set these names will be used to label the
# highlighted wind sectors.
foehnix:::windsector_convert(matrix(c(10, 30, 90, 140), byrow = TRUE, ncol = 2,
                         dimnames = list(c("A", "B"), c("from", "to"))))
foehnix:::windsector_convert(list(A = c(10, 30), B = c(90, 140)))
foehnix:::windsector_convert(structure(data.frame(from = c(30, 90), to = c(30, 140)),
                                       row.names = c("foo", "bar")))

}
\author{
Reto Stauffer
}
