% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{write.csv.foehnix}
\alias{write.csv.foehnix}
\title{Write Estimated Probabilities to CSV File}
\usage{
\method{write.csv}{foehnix}(x, file, info = TRUE, format = NULL, ...)
}
\arguments{
\item{x}{a \code{\link[foehnix]{foehnix}} object}

\item{file}{character, name of the target file}

\item{info}{logical, whether or not header information
should be written}

\item{format}{NULL or a character to specify the format in which the
date/time information should be written to the \code{file} (forwarded to
\code{strftime})}

\item{...}{currently ignored}
}
\value{
Invisible return of the data.frame written to the output file.
}
\description{
Write the results of a \code{\link[foehnix]{foehnix}} model into a CSV
text file. Custom date/time information can be specified using the
input argument \code{format}. By default UNIX timestamp will be used.
}
\author{
Reto Stauffer
}
