% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{write.csv}
\alias{write.csv}
\title{Data Output}
\usage{
write.csv(...)
}
\arguments{
\item{...}{arguments passed to generic method.}
}
\description{
Generic method to write data to CSV. By default this method
falls back to \code{utils::write.csv(...)}.
}
\seealso{
foehnix()
}
