% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{coef.foehnix}
\alias{coef.foehnix}
\title{Get Estimated Mixture Model Coefficients}
\usage{
\method{coef}{foehnix}(object, type = "parameter", ...)
}
\arguments{
\item{object}{a \code{\link[foehnix]{foehnix}} object.}

\item{type}{character, either \code{"parameter"} \code{"coefficients"}, see
'Details' section.}

\item{...}{additional arguments, ignored.}
}
\value{
Returns a \code{coef.foehnix} object with the estimated
coefficients.
}
\description{
Returns the estimated coefficients of a \code{\link[foehnix]{foehnix}}
mixture model for both, the components and the concomitant model (if specified).
}
\details{
Returns the coefficients of the mixture model.  If \code{type =
"parameter"} the parameters of the mixture model components are returned on
the 'true' scale (parameter scale), else on the link scale (as used for
optimization).  For each component the location and scale parameters are
returned.  If a concomitant model has been specified, the coefficients of
the (possibly regularized) logistic regression model are returned as well
(concomitant model).
}
\seealso{
\code{\link[foehnix]{foehnix}}, \code{\link[foehnix]{iwls_logit}}.
}
\author{
Reto Stauffer
}
