% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions.R
\name{ddff2uv}
\alias{ddff2uv}
\title{Conver Wind Speed and Wind Direction In U/V Components}
\usage{
ddff2uv(dd, ff = NULL)
}
\arguments{
\item{dd}{only necessary if 'ff' is a numeric vector.
Use NULL if ff is containing both
variables (ff/dd), else dd is a data vector containing
the wind direction in degrees (0: north, 90: east, ...)}

\item{ff}{there are several options to use this input argument.
Most simple: ff is a numeric vector, then dd is required as
second input (numeric vector, too). But ff can also be
a zoo object or a data.frame or a matrix.
If it is a matrix: needs two columns, first one has to be
ff, second dd. If it is a zoo object or a data.frame
there have to be at least two columns (only those two will
be used for computation) named 'ff' and 'dd'.}
}
\value{
Returns a data.frame containing the \code{u} and \code{v} components
        of the data. In addition, \code{rad} (mathematical representation
        of wind direction in radiant) is returned.
}
\description{
Converting wind direction (dd) and wind speed (ff) into 
zonal (u) und meridional (v) wind components.
}
\details{
Wind speed units can be what you want. You are getting back
         the two wind components in the same unit. Wind direction 'dd'
         has to be in meteorological degrees where 0 is North,
         90 is East and so on.
}
\examples{
## Generate dd and ff variable
set.seed(0)
ff <- floor(abs(rnorm(20))*10)
dd <- sample(seq(0,359),20)
df <- data.frame('ff'=ff,'dd'=dd)

## Using with vectors 
print(head(ddff2uv( 'ff' = ff, 'dd' = dd )))

## Using with data.frame
print(head(ddff2uv(df)))

## Using with matrix
print(head(ddff2uv(as.matrix(df))))

}
\seealso{
uv2ddff
}
\author{
Jakob Messner
}
