% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{demodata}
\alias{demodata}
\title{foehnix Demo Data Set}
\source{
Station operated by the Department of Atmospheric and Cryospheric
Sciences (\url{http://acinn.uibk.ac.at}) of the University of Innsbruck.
The data is available under the
\href{https://creativecommons.org/licenses/by-sa/4.0/}{CC BY-SA 4.0 license}.
}
\usage{
demodata(which = "combined")
}
\arguments{
\item{which}{either \code{"combined"} (returns a combined data set),
\code{"sattelberg"} (only observations from station Sattelberg),
or \code{"ellboegen"} (only observations rom station Ellboegen).}
}
\description{
The foehnix package comes with two demo data sets containing
observations from two automated weather stations in Tyrol,
Austria. One station (Sattelberg) is located close to the main
alpine ridge and is used as 'crest' station. The second station
(Ellboegen) is located in the Wipp valley.
The package demos and examples will estimate automated foehn
classification models (\code{\link{foehnix}} models) for 
station Ellboegen using additional information from the
crest station as concomitant variables and for custom wind
filters (\code{\link{foehnix_filter}}).
}
\examples{
# Loading the combined demo data set. Variables starting with
# "crest_" are observations from station Sattelberg.
x <- demodata()
print(head(x))

# Sattelberg only
x <- demodata("sattelberg")
print(head(x))

# Ellboegen only
x <- demodata("ellboegen")
print(head(x))

}
\seealso{
\code{\link{ellboegen}}, \code{\link{sattelberg}}
}
\author{
Reto Stauffer
}
