% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/demodata.R
\docType{data}
\name{demodata}
\alias{demodata}
\alias{ellboegen}
\alias{sattelberg}
\alias{viejas}
\alias{luckyfive}
\title{foehnix Demo Data Set}
\source{
Station Ellögen operated by the Department of Atmospheric and Cryospheric
Sciences (\url{http://acinn.uibk.ac.at}) of the University of Innsbruck.
The data is available under the
\href{https://creativecommons.org/licenses/by-sa/4.0/}{CC BY-SA 4.0 license}.

Station Sattelberg operated by the Department of Atmospheric and Cryospheric
Sciences (\url{http://acinn.uibk.ac.at}) of the University of Innsbruck.
The data is available under the
\href{https://creativecommons.org/licenses/by-sa/4.0/}{CC BY-SA 4.0 license}.

Station Viejas Casino and Resort operated by San Diego Gas and Electric and made freely
publicly available through Synoptic PBC at \url{https://synopticdata.com/}.

Station Lucky Five Ranch operated by San Diego Gas and Electric and made freely
publicly available through Synoptic PBC at \url{https://synopticdata.com/}.
}
\usage{
demodata(which = c("tyrol", "ellboegen", "sattelberg", "california",
  "viejas", "luckyfive"))
}
\arguments{
\item{which}{\code{which = "tyrol"} returns the combined tyrolean data set (A) and
\code{which = "california"} the combined californian data set (USA), a time series
object which contains measurements from both stations (valley/crest).
The function can also be used to get the demo data sets for specific sites
(see 'Examples'/'Usage').}
}
\description{
The \code{\link{foehnix}} package comes with two sets of meteorological
observations one for Tyrol, Austria, and one for Southern California, USA.
Each region comes with observations from two stations, one valley station
(or target station) and one station further upstream of the main wind direction
(crest station) used to filter the data (see \code{\link{foehnix_filter}}).
For Tyrol, observations for station Ellbögen (valley) and station Sattelberg
(crest) are included, the Californian data set consists of the crest station
'Lucky Five Ranch' and the valley station 'Viejas Casino and Resort'.
More details are provided in the 'Details' section.
}
\details{
The Tyrolean data set consists of station
Elbögen located in the Wipp valley, an area well known for south foehn
events. The second station called Sattelberg is
located 18km (11.5mi) south of Ellbögen close to the main
alpine ridge and is used as upstream crest station.

\itemize{
    \item Ellbögen: 11.42889 E/47.18694 N, 1080 meters above mean sea level
    \item Sattelberg: 11.47889 E/47.01083 N, 2107 meters above mean sea level
}

Hourly time series objects (class `zoo`; UTC) including the following variables:
\itemize{
   \item \code{dd}: wind direction in degrees, meteorological format
       (0/360 correspond to 'wind from north', 90 to 'wind from east',
       180 'wind from south', and 270 'wind from west')
   \item \code{ff}: wind speed in meters per second
   \item \code{p}: station pressure in hectopascal
   \item \code{rh}: relative humidity in percent
   \item \code{t}: dry air temperature in degrees Celsius
}

The Californian data sets consist of two stations located in the southern
part of the state.  Station 'Viejas' is installed next to the Viejas Casino
and Resort located in Alpine, CA (valley station), the station 'Lucky Five
Ranch' is located 22km (14mi) northwest of the Viejas Casino close to the
main ridge of the Sierra Nevada mountain range and is used as the upstream
crest station.

\itemize{
    \item Viejas Casino and Resort: -116.70437 E/32.84559 N, 715 meters above mean sea level
    \item Lucky Five Ranch: -116.528 E/32.9331 N, 1445 meters above mean sea level
}

Hourly time series objects (class `zoo`; UTC) including the following variables:
\itemize{
   \item \code{air_temp}: dry air temperature in degrees Celsius
   \item \code{relative_humidity}: relative humidity in percent
   \item \code{wind_speed}: wind speed in meters per second
   \item \code{wind_direction}: wind direction in degrees, meteorological format
       (0/360 correspond to 'wind from north', 90 to 'wind from east',
       180 'wind from south', and 270 'wind from west')
   \item \code{wind_gust}: wind gust speed in meters per second
}
}
\examples{
# Loading the combined demo data set for "Tyrol (A)".
# Stations: Ellboegen (valley station) and Sattelberg (crest station).
# Variables starting with "crest_" are observations from station Sattelberg,
x <- demodata("tyrol") # Default
print(head(x))

# Loading the combined demo data set for "California (USA)".
# Stations: Viejas (valley station) and 'Lucky Five Ranch' (crest station).
# Variables starting with "crest_" are observations from station 'Lucky Five Ranch'.
x <- demodata("california")
print(head(x))

# Sattelberg only
x <- demodata("sattelberg")
print(head(x))

# Solely Ellboegen
x <- demodata("ellboegen")
print(head(x))

# Viejas
x <- demodata("viejas")
print(head(x))

# Lucky Five Ranch
x <- demodata("luckyfive")
print(head(x))

}
\seealso{
\code{\link{ellboegen}}, \code{\link{sattelberg}},
         \code{\link{viejas}}, \code{\link{luckyfive}}
}
\author{
Reto Stauffer, Deborah Detka

Reto Stauffer
}
\keyword{data}
