% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{destandardize}
\alias{destandardize}
\title{Destandardize a Standardized Matrix}
\usage{
destandardize(x, ...)
}
\arguments{
\item{x}{standardized matrix of class \code{standardized}.}

\item{...}{forwarded, unused.}
}
\description{
Destandardize a Standardized Matrix
}
\details{
Reverse function of \code{\link[foehnix]{standardize}}.
}
\examples{
# See example on R documentation for \code{?standardize}.

}
\seealso{
\code{\link[foehnix]{standardize}}
}
\author{
Reto Stauffer
}
