% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix.R, R/foehnix_functions.R
\name{foehnix}
\alias{foehnix}
\alias{logLik.foehnix}
\alias{logLik}
\alias{nobs.foehnix}
\alias{nobs}
\alias{AIC.foehnix}
\alias{AIC}
\alias{BIC.foehnix}
\alias{BIC}
\alias{IGN.foehnix}
\alias{edf.foehnix}
\alias{print.foehnix}
\alias{formula.foehnix}
\alias{summary.foehnix}
\title{Foehn Classification Based on a Two-Component Mixture Model}
\usage{
foehnix(
  formula,
  data,
  switch = FALSE,
  filter = NULL,
  family = "gaussian",
  control = foehnix.control(family, switch, ...),
  ...
)

\method{logLik}{foehnix}(object, ...)

\method{nobs}{foehnix}(object, ...)

\method{AIC}{foehnix}(object, ...)

\method{BIC}{foehnix}(object, ...)

\method{IGN}{foehnix}(object, ...)

\method{edf}{foehnix}(object, ...)

\method{print}{foehnix}(x, ...)

\method{formula}{foehnix}(x, ...)

\method{summary}{foehnix}(object, eps = 1e-04, detailed = FALSE, ...)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be coerced
to that class): a symbolic description of the model to be fitted.  The
details of model specification are given under 'Details'.}

\item{data}{a regular (not necessarily strictly regular)
time series object of class \code{zoo} containing the
variables for the two-part mixture model.}

\item{switch}{logical. If set to \code{TRUE} the two estimated components
will be switched. This is important if the covariate for the components
(left hand side of input \code{formula} is smaller for cases with foehn than
for cases without foehn, e.g., when using the temperature difference as main
covariate.}

\item{filter}{a named list can be provided to apply a custom (simple) filter
to the observations on \code{data}. Can be used to e.g., prespecify a
specific wind sector for foehn winds. Please read the manual of the
\code{\link[foehnix]{foehnix_filter}} method for more details and examples.}

\item{family}{character (at the moment \code{"gaussian"} or \code{"logistic"})
or an object of class \code{foehnix.family}.}

\item{control}{additional control arguments, see \code{\link[foehnix]{foehnix.control}}.}

\item{...}{forwarded to \code{\link[foehnix]{foehnix.control}}}

\item{object}{a \code{foehnix} object (input for S3 methods)}

\item{x}{a \code{foehnix} object (input for S3 methods)}

\item{eps}{threshold for posterior probabilities used in \code{summary}.
\code{numeric} value, default \code{1e-4}.}

\item{detailed}{boolean, default \code{FALSE}. If \code{TRUE}, additional information
will be returned.}
}
\value{
Returns an object of class \code{foehnix}.

Log-likelihood sum (\code{numeric}).

Number of observations (\code{integer}) used to train the model.

Returns the Akaike information criterion (\code{numeric}).

Bayesian information criterion (\code{numeric}).

Ignorance (mean negative log-likelihood; \code{numeric}).

Effective degrees of freedom.

Returns the formula of the model (\code{formula}).

Object of class \code{summary.foehnix}.
}
\description{
This is the main method of the foehnix package to estimate
two-component mixture models for automated foehn classification.
}
\details{
The two-component mixture model can be specified via formula object where
the left hand side of the formula contains the 'main' variable explaining the two
components (only one variable), the right hand side of the formula specifies
the concomitant variables (multiple variables allowed).  As an example:
let's assume that our zoo object \code{data} contains the following
columns:
\itemize{
    \item \code{ff}: observed wind speed at target site
    \item \code{rh}: observed relative humidity at target site
    \item \code{diff_t}: (potential) temperature difference between target site and a station
          upstream of the foehn wind direction
}

The specification for \code{formula} could e.g. look as follows:
\itemize{
    \item \code{ff ~ 1}: the two components of the mixture model will be
          based on the observed wind speed (\code{ff}), no concomitant
          model (right hand side is simply \code{1}). 
    \item \code{ff ~ rh}: similar to the specification above, but using
           observed relative humidity (\code{rh}) as concomitant.
    \item \code{ff ~ rh + diff_t}: as above but with an additional second
          concomitant variable (observed temperature difference, \code{diff_t}).
    \item \code{diff_t ~ ff + rh}: using temperature difference as the main
          variable for the two components while \code{ff} and \code{rh} are
          used as concomitants. Note: in this case it will be required to set
          \code{switch = TRUE} as lower values of \code{diff_t} indicate
          less stratified conditions where the occurrence of foehn is
          more likely. If \code{switch = FALSE} the two components (foehn
          and no foehn) may be interchanged and the model will return
          "probabilities of not observing foehn".
}
 
Note that these are just examples and have to be adjusted given data
availability, location, structure/names of the variables in the \code{data}
object.

The optional input \code{filter} allows to specify simple or complex
filters (see \code{\link[foehnix]{foehnix_filter}} for details).
This allows to add additional constraints, e.g., adding a filter on the
observed wind direction to ensure that only events within a specific wind
sector (the main foehn wind direction) are classified as "foehn".
}
\references{
Plavcan D, Mayr GJ, Zeileis A (2014).
Automatic and Probabilistic Foehn Diagnosis with a Statistical Mixture Model.
\emph{Journal of Applied Meteorology and Climatology}.
\bold{53}(3), 652--659. \doi{10.1175/JAMC-D-13-0267.1}

Gr\\"un B, Leisch F (2007).
Fitting Finite Mixtures of Generalized Linear Regressions in \emph{R}.
\emph{Computational Statistics \& Data Analysis}. \bold{51}(11), 5247--5252.
\doi{10.1016/j.csda.2006.08.014}

Gr\\"un B, Leisch F (2008).
FlexMix Version 2: Finite Mixtures with Concomitant Variables and Varying and Constant Parameters.
\emph{Journal of Statistical Software, Articles}.
\bold{28}(4), 1--35. \doi{10.18637/jss.v028.i04}

Fraley C, Raftery AE (2002).
Model-Based Clustering, Discriminant Analysis, and Density Estimation.
\emph{Journal of the American Statistical Association}.
\bold{97}(458), 611--631. \doi{10.1198/016214502760047131}
}
\seealso{
See \code{\link[foehnix]{foehnix_filter}} for more information about the
\code{filter} option.  See also: \code{\link[foehnix]{tsplot}}, \code{\link[foehnix]{windrose}}.

Foehnix family objects: \code{\link[foehnix]{foehnix.family}}.

S3 methods for \code{foehnix} objects:
\code{\link[foehnix]{plot.foehnix}},
\code{\link[foehnix]{predict.foehnix}},
\code{\link[foehnix]{fitted.foehnix}},
\code{\link[foehnix]{print.foehnix}},
\code{\link[foehnix]{summary.foehnix}},
\code{\link[foehnix]{windrose.foehnix}},
\code{\link[foehnix]{coef.foehnix}},
\code{\link[foehnix]{nobs.foehnix}},
\code{\link[foehnix]{edf.foehnix}},
\code{\link[foehnix]{AIC.foehnix}},
\code{\link[foehnix]{BIC.foehnix}},
\code{\link[foehnix]{IGN.foehnix}},
\code{\link[foehnix]{logLik.foehnix}},
...
}
\author{
Reto Stauffer
}
