% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix.R, R/foehnix_functions.R
\name{foehnix}
\alias{foehnix}
\alias{nobs.foehnix}
\alias{logLik.foehnix}
\alias{AIC.foehnix}
\alias{BIC.foehnix}
\alias{edf.foehnix}
\alias{print.foehnix}
\alias{formula.foehnix}
\alias{summary.foehnix}
\alias{image.foehnix}
\title{Foehn Classification Based on a Two-Component Mixture Model}
\usage{
foehnix(formula, data, switch = FALSE, filter = NULL,
  family = "gaussian", control = foehnix.control(family, switch, ...),
  ...)

\method{nobs}{foehnix}(object, ...)

\method{logLik}{foehnix}(object, ...)

\method{AIC}{foehnix}(object, ...)

\method{BIC}{foehnix}(object, ...)

\method{edf}{foehnix}(object, ...)

\method{print}{foehnix}(x, ...)

\method{formula}{foehnix}(x, ...)

\method{summary}{foehnix}(object, detailed = FALSE, ...)

\method{image}{foehnix}(x, FUN = "freq", deltat = NULL, deltad = 7L,
  col = rev(gray.colors(20)), contours = FALSE,
  contour.col = "black", ...)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be coerced
to that class): a symbolic description of the model to be fitted.  The
details of model specification are given under 'Details'.}

\item{data}{a regular (not necessarily strictly regular)
time series object of class \code{zoo} containing the
variables for the two-part mixture model.}

\item{switch}{logical. If set to \code{TRUE} the two estimated components
will be switched. This is important if the covariate for the components
(left hand side of input \code{formula} is smaller for cases with foehn than
for cases without foehn, e.g., when using the temperature difference as main
covariate.}

\item{filter}{a named list can be provided to apply a custom (simple) filter
to the observations on \code{data}. Can be used to e.g., prespecify a
specific wind sector for foehn winds. Please read the manual of the
\code{\link{foehnix_filter}} method for more details and examples.}

\item{family}{character (at the moment \code{"gaussian"} or \code{"logistic"})
or an object of class \code{foehnix.family}.}

\item{control}{additional control arguments, see \code{\link{foehnix.control}}.}

\item{...}{forwarded to \code{\link{foehnix.control}}}

\item{object}{a \code{foehnix} object (input to S3 methods)}

\item{x}{object of class \code{\link{foehnix}}.}

\item{detailed}{boolean, default FALSE. If set to true, additional information
will be printed.}

\item{FUN}{character string or a custom aggregation function. See 'Details'
section for more information.}

\item{deltat}{integer, interval in seconds for the time of day axis. Has to be
a fraction of 86400 (24 hours in seconds). It \code{NuLL} (default) the
interval of the time series object will be used.}

\item{deltad}{integer, similar to \code{deltat}, the interval in days for the
grid on the x-axis. Default is \code{7L} (aggregate to weekly values).}

\item{col}{vector of colors forwarded to \code{image.default}. By default
a gray scale color map is used.}

\item{contours}{logical \code{TRUE} or \code{FALSE}, whether or not the
concours should be added.}

\item{contour.col}{color for the contour lines, only used if \code{contours = TRUE}.}
}
\value{
Returns an object of class \code{foehnix}.
}
\description{
This is the main method of the foehnix package to estimate
two-component mixture models for automated foehn classification.

The default \code{\link{image}} plot of a \code{\link{foehnix}} object
is a Hovmoeller diagram.
}
\details{
The two-component mixture model can be specified via formula object where
the left hand side of the formula contains the variable explaining the two
components (only one variable), the right hand side of the formula specifies
the concomitant variables (multiple variables allowed).  As an example:
let's assume that our zoo object '\code{data}' contains the following
columns:
\itemize{
    \item \code{ff}: observed wind speed at target site
    \item \code{rh}: observed relative humidity at target site
    \item \code{diff_t}: temperature difference between target site and a station
          upstream of the foehn wind direction
}

The specification for \code{formula} could e.g. look as follows:
\itemize{
    \item \code{ff ~ 1}: the two components of the mixture model will be
          based on the observed wind speed (\code{ff}), no concomitant
          variables are given (intercept only). 
    \item \code{ff ~ rh}: similar to the specification above, but using
           observed relative humidity (\code{rh}) as concomitant.
    \item \code{ff ~ rh + diff_t}: as above but with an additional second
          concomitant variable (observed temperature difference, \code{diff_t}).
    \item \code{diff_t ~ ff + rh}: using temperature difference as the main
          variable for the two components while \code{ff} and \code{rh} are
          used as concomitants.
}
 
Note that these are just examples and have to be adjusted given data
availability, location, structure/names of the variables in the \code{data}
object.

An optional numeric vector of length \code{2} can be provided to filter the
data. If set, only times (rows in \code{data}) where the observed wind
direction \code{dd} lies within the wind sector specified.  For the examples
the wind direction \code{dd} is assumed to be in degrees, meteorological
definition (\code{90} = from east, \code{180} = from south, \dots).  The
following condition is used:
\itemize{
    \item If \code{windsector[1L] < windsector[2L]}: use \code{data} where
          \code{data$dd >= windsector[1L]} \code{&} \code{data$dd <= windsector[2L]}
          (e.g., \code{windsector = c(90, 270)} for south wind sector).
    \item If \code{windsector[1L] > windsector[2L]}: use \code{data} where
          \code{data$dd >= windsector[1L]} \code{|} \code{data$dd <= windsector[2L]}
          (e.g., \code{windsector = c(270, 90)} for north wind sector).
}

Plotting a Hovmoeller diagram based on the \code{\link{zoo}} time
series object of the \code{\link{foehnix}} classification. Different plot
types are available. The default functions (see list below) use \code{na.rm = TRUE}.

Input \code{FUN} can be one of the following character strings:
\itemize{
   \item \code{mean}: mean probability.
   \item \code{freq}: plotting frequencies.
   \item \code{occ}: plotting occurance of foehn (where probability \code{>= 0.5}).
   \item \code{noocc}: contrary to \code{occ}: occurance of no foehn (probability \code{< 0.5}).
}

\code{FUN} can also be a custom function used for time series aggregation
(see \code{\link{aggregate.zoo}}).
}
\references{
Plavcan D, Mayr GJ, Zeileis A (2014).
Automatic and Probabilistic Foehn Diagnosis with a Statistical Mixture Model.
\emph{Journal of Applied Meteorology and Climatology}.
\bold{53}(3), 652--659. \doi{10.1175/JAMC-D-13-0267.1}

Gr\\"un B, Leisch F (2007).
Fitting Finite Mixtures of Generalized Linear Regressions in \emph{R}.
\emph{Computational Statistics \& Data Analysis}. \bold{51}(11), 5247--5252.
\doi{10.1016/j.csda.2006.08.014}

Gr\\"un B, Leisch F (2008).
FlexMix Version 2: Finite Mixtures with Concomitant Variables and Varying and Constant Parameters.
\emph{Journal of Statistical Software, Articles}.
\bold{28}(4), 1--35. \doi{10.18637/jss.v028.i04}

Fraley C, Raftery AE (2002).
Model-Based Clustering, Discriminant Analysis, and Density Estimation.
\emph{Journal of the American Statistical Association}.
\bold{97}(458), 611--631. \doi{10.1198/016214502760047131}
}
\seealso{
See \code{\link{foehnix_filter}} for more information about the
\code{filter} option.  See also: \code{\link{tsplot}}, \code{\link{windrose}}.
TODO: filter should be renamed to filter.

Foehnix family objects: \code{\link{foehnix.family}}.

S3 methods for \code{foehnix} objects:
\code{\link{plot.foehnix}}, \code{\link{predict.foehnix}}, \code{\link{fitted.foehnix}},
\code{\link{print.foehnix}}, \code{\link{windrose.foehnix}}, \code{\link{coef.foehnix}},
...
}
\author{
Reto Stauffer
}
