% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix.R
\name{foehnix.reg.fit}
\alias{foehnix.reg.fit}
\title{Fitting Regularized foehnix Mixture Model With Concomitant Model.}
\usage{
foehnix.reg.fit(
  y,
  logitX,
  family,
  glmnet.control,
  switch = FALSE,
  maxit = 100L,
  tol = 1e-05,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{y}{numeric vector, covariate for the components of the
mixture model, dimension \code{N}.}

\item{logitX}{numeric matrix of dimension \code{N x P}, covariates
for the concomitant model (logistic regression model matrix).}

\item{family}{object of class \code{\link[foehnix]{foehnix.family}}.}

\item{glmnet.control}{an object of class \code{\link[foehnix]{glmnet.control}}
containing the arguments for the glmnet function.}

\item{switch}{logical whether or not the two components should be switched.
By default (\code{switch = FALSE}) the component which shows
higher values of \code{y} is assumed to be the foehn cluster!
Depending on what your covariate is you might need to switch
the clusters (by setting \code{switch = TRUE}).}

\item{maxit}{positive integer, or vector of length 2 with positive
integer values. Maximum number of iterations of the EM algorithm
and the concomitant model.
Check manual of \code{\link[foehnix]{foehnix.control}} for more details.}

\item{tol}{numeric, or vector of length 2 containing numeric values.
Tolerance for the optimization (EM algorithm and concomitant model).
Check manual of \code{\link[foehnix]{foehnix.control}} for more details.}

\item{verbose}{logical, default \code{TRUE}. If set to \code{FALSE}
verbose output will be suppressed.}

\item{...}{currently unused.}
}
\description{
Fitting method for two-component foehnix mixture model with regularization.
Typically not called directly, interfaced via \code{\link[foehnix]{foehnix}}.
}
\details{
TODO: Method not yet implemented, as soon as the method itself
has been written: update manual page!
}
\seealso{
\code{\link[foehnix]{foehnix}}, \code{\link[foehnix]{foehnix.control}},
\code{\link[foehnix]{foehnix.noconcomitant.fit}}, \code{\link[foehnix]{foehnix.reg.fit}},
\code{\link[foehnix]{iwls_logit}}.
}
\author{
Reto Stauffer
}
