% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{IGN}
\alias{IGN}
\title{Returns the Ignorance of an Object}
\usage{
IGN(object, ...)
}
\arguments{
\item{object}{the object to be evaluated.}

\item{...}{forwarded to generic methods.}
}
\description{
Returns the ignorance (mean negative log-likelihood)
of the \code{object}.
}
\seealso{
\code{\link[foehnix]{foehnix}}
}
