% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.foehnix.R
\name{plot.foehnix}
\alias{plot.foehnix}
\title{foehnix Model Assessment Plots}
\usage{
\method{plot}{foehnix}(x, which = NULL, log = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{foehnix}} mixture model object.}

\item{which}{\code{NULL} (default), character, character string,
integer, or numeric. Allowed characters: \code{loglik},
\code{loglikcontribution}, \code{coef}, and \code{hist}. If \code{which}
is numeric/integer or a vector of numerics/integers the numbers
correspond to \code{loglik} (\code{1}), \code{loglikpath} (\code{2}),
or \code{coef} (\code{3}).}

\item{log}{logical, if \code{TRUE} the x-axis is shown on the log scale,
else on the iteration scale.}

\item{...}{additional arguments, unused.}
}
\description{
Visual representation \code{\link{foehnix}} model optimization.
}
\details{
There are currently three different plot types.
\itemize{
    \item \code{"loglik"} shows the log-likelihood sum path trough
        the iterations of the EM algorithm for parameter estimation.
    \item \code{"loglikcontribution"} shows the log-likelihood
        contribution (initial value subtracted; all paths start
        with \code{0}).
    \item \code{coef} shows the development of the (standardized)
        coefficients during EM optimization. Parameters of the
        components are shown on the real scale, the coefficients
        of the concomitant model (if used) are shown on the
        standardized scale.
}
}
\author{
Reto Stauffer
}
