% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foehnix_functions.R
\name{standardize}
\alias{standardize}
\alias{standardize.matrix}
\alias{scale.standardized}
\alias{center}
\alias{destandardize_coefficients}
\title{Standardize (Model) Matrix}
\usage{
standardize(x, ...)

\method{standardize}{matrix}(x, ...)

\method{scale}{standardized}(x, ...)

center(x, ...)

destandardize_coefficients(beta, X)
}
\arguments{
\item{x}{matrix of dimension \code{N x p}.}

\item{...}{additional arguments, ignored.}

\item{beta}{regression coefficients estimated on standardized data.}

\item{X}{object of class \code{\link{standardize}}.}
}
\value{
Returns a matrix of the same dimension as input \code{x}
but with standardized data. The return object is of class
\code{c("standardized", "matrix")} which comes with some handy
S3 methods.

Returns destandardized regression coefficients, same object
as input \code{beta}.
}
\description{
Function to standardize the columns of a matrix, used to
standardize the model matrix before estimating the regression
coefficients of a generalized linear model (\code{\link{iwls_logit}}).

Destandardize coefficients. Brings coefficients back to
the "real" scale if standardized coefficients are used when
estimating the logistic regression model (concomitant model).
}
\examples{
# Example data set
data("airquality")
airquality <- na.omit(airquality)

# Create model matrix
X <- model.matrix(Ozone ~ ., data = airquality)
print(head(X))

# Standardize
S <- standardize(X)
print(head(S))

is.standardized(X)
is.standardized(X)

# Get parameters used for standardization
center(S)
scale(S)

}
\seealso{
\code{\link{standardize}}. Used in \code{\link{foehnix}}
and \code{\link{iwls_logit}}.
}
\author{
Reto Stauffer
}
