% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\title{Time Series Plot of foehnix Models}
\usage{
tsplot(x, start = NULL, end = NULL, ndays = 10,
  control = tsplot.control(...), ..., ask = TRUE)
}
\arguments{
\item{x}{object of type \code{foehnix} or a list with \code{foehnix} and
univariate \code{zoo} objects, see 'Details'.}

\item{start}{POSIXt object or an object which can be converted to POSIXt.}

\item{end}{POSIXt object or an object which can be converted to POSIXt.}

\item{ndays}{integer, number of days used when looping trough the time series.}

\item{control}{an object of class \code{\link[foehnix]{tsplot.control}}.}

\item{...}{additional arguments forwarded to \code{\link[foehnix]{tsplot.control}}.
can be used to rename varaiables and to change the look of the time series
plot. Please see 'Examples' and the manual of the function
\code{\link[foehnix]{tsplot.control}} for more details.}

\item{ask}{logical, default is \code{TRUE}.}
}
\description{
Development time series plots of estimated \code{\link{foehnix}}
models. TODO: they are very specific at the moment!
}
\details{
Development method to access plausability of the estimated foehn
proabilities.  For software release this method should either be removed or
made much more general. At the moment the method heavily depends on the
names of the data used as input for \code{\link{foehnix}}.

This time series plotting function creates a relatively specific plot
and also expects, by default, a set of default variables and variable
names. This function uses the data set provided on the \code{data}
input argument when calling the \code{\link{foehnix}} method.
As they might differ from the \code{foehnix} defaults the
\code{varnames} input argument allows to specify custom names.
The ones expected:

\itemize{
     \item \code{t}: dry air temperature
     \item \code{rh}: relative humidity (in percent)
     \item \code{diff_t}: temperature difference between the
           crest and the valley station
     \item \code{dd}: meteorological wind direction (\code{[0, 360]})
     \item \code{ff}: wind speed
}

Custom names can be specified by renaming the defaults based on a
named list. As an example: assume that wind direction is called
\code{winddir} and wind speed is called \code{windspd} in your data set.
Specify the following input to rename/respecify the defaults:
\itemize{
      \item \code{varnames = list(dd = "winddir", ff = "windspd")}
}

Please note: if a variable cannot be found (no matter whether
the default variable names have been renamed or not) this specific
variable will be ignored. Subplots will not be displayed if no
data are available at all.

TODO: describe input 'x'
}
\seealso{
\code{\link[foehnix]{tsplot.control}}.
}
\author{
Reto Stauffer
}
