% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\title{Time Series Plot of foehnix Models}
\usage{
tsplot(x, start = NULL, end = NULL, ndays = 10, ..., xtra = NULL,
  ask = TRUE)
}
\arguments{
\item{x}{object of type \code{foehnix}.}

\item{start}{POSIXt object or an object which can be converted to POSIXt.}

\item{end}{POSIXt object or an object which can be converted to POSIXt.}

\item{ndays}{integer, number of days used when looping trough the time series.}

\item{...}{additional arguments, ignored.}

\item{xtra}{optional zoo object with probabilities in \code{]0,1[} to compare
two classification algorithms.}

\item{ask}{logical, default is \code{TRUE}.}
}
\description{
Development time series plots of estimated \code{\link{foehnix}}
models. TODO: they are very specific at the moment!
}
\details{
Development method to access plausability of the estimated foehn
proabilities.  For software release this method should either be removed or
made much more general. At the moment the method heavily depends on the
names of the data used as input for \code{\link{foehnix}}.

TODO: this is very specific to our data and our variable names.
      either provide something like this and force the users to
      follow our naming conventions, or make it much more
      flexible/generig.
}
\author{
Reto Stauffer
}
