% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\alias{windrose.foehnix}
\title{Windrose Plot}
\usage{
windrose(x, which = NULL, ddvar = "dd", ffvar = "ff", ...)

\method{windrose}{foehnix}(x, which = NULL, ddvar = "dd",
  ffvar = "ff", ...)
}
\arguments{
\item{x}{\code{zoo} object or \code{data.frame} or \code{\link{foehnix}}
mixture model object.}

\item{which}{character, one of \code{"unconditional"}, \code{"nofoehn"}, 
\code{"foehn"}, or \code{"combined"}.}

\item{ddvar}{character, name of the wind direction varaible in \code{x} or as used
for training the \code{\link{foehnix}} mixture model.}

\item{ffvar}{character, name of the wind speed varaible in \code{x} or as used
for training the \code{\link{foehnix}} mixture model.}

\item{...}{additional arguments, currently unused.}
}
\description{
Plotting a windrose using meteorological wind direction and
wind speed (or gust speed).
}
\details{
TODO: not yet finished and/or tested!
}
