% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\alias{windrose.foehnix}
\alias{windrose.default}
\title{Windrose Plot}
\usage{
windrose(x, ...)

\method{windrose}{foehnix}(x, type = NULL, which = NULL,
  ddvar = "dd", ffvar = "ff", mfcol = 2, maxpp = Inf, ...)

\method{windrose}{default}(dd, ff, interval = 10, type = "density",
  windsector = NULL, main = NULL, hue = c(10, 100), power = 0.5)
}
\arguments{
\item{x}{forwarded to windrose methods.}

\item{...}{forwarded as well.}

\item{type}{\code{NULL} or character, one of \code{density} or \code{histogram}.
If \code{NULL} both types will be plotted.}

\item{which}{\code{NULL} or character, one of \code{unconditional}, \code{nofoehn},
or \code{foehn}. If \code{NULL} all three will be plotted.}

\item{ddvar}{character, name of the column in the training data set which contains
the wind direction information.}

\item{ffvar}{character, name of the column in the training data set which contains
the wind speed (or gust speed) data.}

\item{mfcol}{integer, number of columns of subplots.}

\item{maxpp}{integer (\code{>0}), maximum plots per page. Not all plots fit on one
page the script loops trough.}

\item{dd}{zoo object with meteorological wind directions (\code{]0, 360[},
0/360 is 'wind from north', 90 'wind from east', 180 'wind from south',
and 270 'wind from west'.}

\item{ff}{zoo object with sind speed data (\code{>= 0}).}

\item{interval}{numeric, a fraction of 360. If 360 cannot be devided by
\code{interval} without a rest the script will stop. Interval
for the segments along the wind direction \code{dd}. Default \code{10}.}

\item{windsector}{TODO: currently unused.}

\item{main}{character, title. If no title is set a default one will be shown.}

\item{hue}{numeric vector of length 1 or two (for \code{type = "density"}
two are used, for \code{type = "histogram"} only the first one will be
used at the moment). Hue(s) for \code{colorspace::heat_hcl(...)}.}

\item{power}{numeric \code{>0}, power parameter for the alpha channel if 
\code{type = "histogram"}. If \code{1} the alpha channel is linear
for the whole range of \code{ff}, values \code{!= 1} change the
alpha behaviour.}

\item{x}{object of class \code{\link{foehnix}}.}

\item{type}{character, one of \code{density} or \code{histogram}, controls
the plot type.}
}
\description{
Plotting a windrose using meteorological wind direction and
wind speed (or gust speed).

Windrose plot based on a \code{\link{foehnix}} mixture model object.

Windrose plot, can handle two type of plots (see \code{type}) using
the same information.
}
\details{
Allows to draw windrose plots from a \code{\link{foehnix}} mixture model
object. Two \code{type}s are available: circular density plots and circular
histograms, both for \code{which = "unconditional"}, \code{which = "nofoehn"},
and \code{which = "foehn"}, details below.
By default (\code{type = NULL}, \code{which = NULL}) all combinations will be plotted.

\itemize{
   \item \code{which = "unconditional"}: unconditional windrose (all observations
         of the data set used to estimate the \code{\link{foehnix}} model).
   \item \code{which = "nofoehn"}: windrose of all observations which have not been
         classified as foehn (probability \code{< 0.5}).
   \item \code{which = "foehn"}: windrose of all observations classified as foehn events
         (probability \code{>= 0.5}).
}

Specific combinations can be selected by calling the windrose function with e.g.,
\code{type = "histogram"} and \code{which = c("foehn", "nofoehn")} (will show
histograms for foehn and no foehn events), or \code{type = NULL} and \code{which = "foehn"}
(will show density and histogram plot for foehn events).
}
\seealso{
windrose.default windrose.foehnix
}
\author{
Reto Stauffer

Reto Stauffer
}
