% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\alias{windrose.foehnix}
\title{Windrose Plot for foehnix Mixture Models and Observations}
\usage{
windrose(x, ...)

\method{windrose}{foehnix}(x, type = NULL, which = NULL,
  ddvar = "dd", ffvar = "ff", breaks = NULL, mfcol = 2L,
  maxpp = Inf, ...)
}
\arguments{
\item{x}{object of class \code{\link{foehnix}} if the windrose plot is applied
to a model, or a vector of wind directions for \code{windrose.default}
(see 'Details' section).}

\item{...}{forwarded to \code{\link{windrose.default}}.}

\item{type}{\code{NULL} or character, one of \code{density} or \code{histogram}.
If \code{NULL} both types will be plotted.}

\item{which}{\code{NULL} or character, one of \code{unconditional}, \code{nofoehn},
or \code{foehn}. If \code{NULL} all three will be plotted.}

\item{ddvar}{character, name of the column in the training data set which contains
the wind direction information.}

\item{ffvar}{character, name of the column in the training data set which contains
the wind speed (or gust speed) data.}

\item{breaks}{\code{NULL} or a numeric vector to define the wind speed (\code{ff}) breaks.}

\item{mfcol}{integer, number of columns of subplots.}

\item{maxpp}{integer (\code{>0}), maximum plots per page. Not all plots fit on one
page the script loops trough.}
}
\description{
foehnix Mixture Model Windrose Plot
}
\details{
Windrose plot based on a \code{\link{foehnix}} mixture model object.

Allows to draw windrose plots from \code{\link{foehnix}} mixture model
object or a set of observations. If input \code{x} to \code{\link{windrose}} is
an object of class \code{foehnix} (as returned by \code{\link{foehnix}}) the
data set which the classification is based on is used to plot the windrose.
If inputs \code{dd} and \code{ff} are given (univariate \code{zoo} time series
objects or \code{numeric} vectors) windrose plots can be plotted for observations
without the need of a \code{foehnix} object.

Two \code{type}s are available: circular density plots and circular
histograms. If the input argument \code{x} is a \code{foehnix} object an additional
input argument \code{which} is available to specify the subset which should be used
to create the windrose plots. The following inputs are allowed:

\itemize{
   \item \code{which = "unconditional"}: unconditional windrose (all observations
         of the data set used to estimate the \code{\link{foehnix}} model).
   \item \code{which = "nofoehn"}: windrose of all observations which have not been
         classified as foehn (probability \code{< 0.5}).
   \item \code{which = "foehn"}: windrose of all observations classified as foehn events
         (probability \code{>= 0.5}).
   \item \code{which = NULL}: all three subsets will be plotted (individual
         windroses).
}
Specific combinations can be specified by calling the windrose function with e.g.,
\code{type = "histogram"} and \code{which = c("foehn", "nofoehn")} (will show
histograms for foehn and no foehn events), or \code{type = NULL} and \code{which = "foehn"}
(will show density and histogram plot for foehn events).
By default (\code{type = NULL}, \code{which = NULL}) all combinations will be plotted.

If \code{dd} and \code{ff} are set only the \code{type} argument is available
(\code{type = "histogram"} or \code{type = "density"}).
}
\examples{
# Loading combined demo data set
data <- demodata("tyrol") # default

# Before estimating a model: plot a wind rose for all observations
windrose(data$dd, data$ff, type = "histogram")
windrose(data$dd, data$ff, type = "density")

# Estimate a foehnix foehn classification model
filter <- list(dd = c(43, 223), crest_dd = c(90, 270)) 
mod <- foehnix(diff_t ~ ff + rh, data = data,
               filter = filter, verbose = FALSE)

# Plotting wind roses
windrose(mod)

# Only density windrose for foehn events
windrose(mod, type = "density", which = "foehn")

# Using custom names: by default wind direction is expected
# to be called 'dd', wind speed is expected to be called 'ff'.
# However, ddvar and ffvar allow to change that (only if
# windrose is called with a foehnix object as input).
# An example:
# - make a copy of data to data2
# - rename dd to winddir, crest_dd to crest_winddir
# - estimate the same foehnix model as above using the
#   new variable names
# - plot windrose with custom names for wind direction (winddir)
#   and wind speed (windspd).

data2 <- data
names(data2)[which(names(data2) == "ff")]       <- "windspd"
names(data2)[which(names(data2) == "dd")]       <- "winddir"
names(data2)[which(names(data2) == "crest_dd")] <- "crest_winddir"
print(head(data2))

filter2 <- list(winddir = c(43, 223), crest_winddir = c(90, 270)) 
mod2 <- foehnix(diff_t ~ windspd + rh, data = data2,
                filter = filter2, verbose = FALSE)

windrose(mod2, type = "density", which = "foehn",
         ddvar = "winddir", ffvar = "windspd")

}
\author{
Reto Stauffer
}
